refactor a bit for navigability
